move tests to better location
